pyproject.toml
[project]
name = "fastapi-opinionated-chat-realtime"
version = "0.1.0"
description = ""
authors = [
    {name = "Azzarnuji",email = "azzarnuzy93@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-opinionated-core @ git+https://github.com/Azzarnuji/fastapi-opinionated-core.git"
]
packages = [
    { include = "fastapi_opinionated_chat_realtime" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]

[tool.poetry.group.dev.dependencies]
twine = "^6.2.0"

